backend/requirements.txt
# PAIfilter Backend Dependencies

# Web Framework
fastapi>=0.127.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI & Vector DB
google-generativeai>=0.8.0
pinecone>=8.0.0

# Utilities
python-dotenv>=1.0.0
